+++ src/ui/main/disable_telemetry.rs
@@ -1,4 +1,5 @@
 use std::process::Command;
+use std::path::PathBuf;
 
 use relm4::prelude::*;
 
@@ -19,11 +20,25 @@ pub fn disable_telemetry(sender: ComponentSender<App>) {
             .collect::<Vec<String>>()
             .join(" ; ");
 
-        let output = Command::new("pkexec")
-            .arg("bash")
-            .arg("-c")
-            .arg(format!("echo '' >> /etc/hosts ; {telemetry} ; echo '' >> /etc/hosts"))
-            .spawn();
+        // TODO: perhaps find some another way? Or doesn't matter?
+        let use_root = std::env::var("LAUNCHER_USE_ROOT")
+            .map(|var| var == "1")
+            .unwrap_or_else(|_| !PathBuf::from("/.flatpak-info").exists());
+
+        let output = if use_root {
+            Command::new("pkexec")
+                .arg("bash")
+                .arg("-c")
+                .arg(format!("echo '' >> /etc/hosts ; {telemetry} ; echo '' >> /etc/hosts"))
+                .spawn()
+        }
+
+        else {
+            Command::new("bash")
+                .arg("-c")
+                .arg(format!("echo '' >> /etc/hosts ; {telemetry} ; echo '' >> /etc/hosts"))
+                .spawn()
+        };
 
         match output.and_then(|child| child.wait_with_output()) {
             Ok(output) => if !output.status.success() {